move DeleteHandler into rest module, fix test fail
